+++ src/main/java/dev/rynwllngtn/agorasystem/entities/comment/Comment.java
@@ -1,6 +1,7 @@
 package dev.rynwllngtn.agorasystem.entities.comment;
 
 import dev.rynwllngtn.agorasystem.dtos.post.AuthorDTO;
+import dev.rynwllngtn.agorasystem.dtos.post.PostDTO;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
@@ -19,6 +20,7 @@ public class Comment {
     @Id
     private String id;
     private AuthorDTO author;
+    private PostDTO post;
     private Date date;
     private String body;